pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentops"
version = "0.0.20"
authors = [
  { name="Alex Reibman", email="alex@agentops.ai" },
  { name="Shawn Qiu", email="shawn@agentops.ai" },
  { name="Braelyn Boynton", email="braelyn@agentops.ai" },
  { name="Howard Gil", email="howard@agentops.ai" }
]
description = "Python SDK for developing AI agent evals and observability"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests==2.31.0",
    "pydantic>=1.9.0",
    "packaging>=23.1,<24.0",
    "toml>=0.10.2",
    "opentelemetry-api>=1.0.0", # TODO: it is recommended that these be optional but the library's main function is observability so makes sense for it to be required. Decide if we need to move
    "opentelemetry-sdk>=1.0.0", # TODO: pick appropriate versions
    "opentelemetry-instrumentation-logging>=0.24b0"
    #"opentelemetry-exporter-otlp>=1.0.0,<2.0.0"

]
[project.optional-dependencies]
dev = [
    "pytest==7.4.0",
    "requests_mock==1.11.0"
]
langchain = [
    "langchain>=0.0.354"
]

[project.urls]
Homepage = "https://github.com/AgentOps-AI/agentops"
Issues = "https://github.com/AgentOps-AI/agentops/issues"
